Move `*-custom.h` headers to `*-custom.h.template`
When using ed25519-donna as a submodule many compilers will look at the
(empty) `*-custom.h` header files which appear first in the directory
include list then not look further for the project-specific headers with
the desired functionality.  This patch fixes the problem by renaming the
stubs to a more appropriate "template" suffix.