Make a new CPP that plays well with the SGX SDK.
As stated in the readme, intel's SGX does not include a `param.h` which
is being included when using the typical set of build flags.  This patch
creates a new CPP, used via `-DSGX_ENCLAVE_ENV`, that explicitly avoids
including `param.h`.

It might be in scope of the donna project (@floodyberry?) to use intel's rng
SHA512 implementation when this flag is set instead of making each dev
create a `-custom.h` header.